cloud/gcp/requirements.txt
pandas==2.1.0
google-cloud-bigquery==3.11.4
tqdm==4.66.1